Added Endpoint type name example name
I was a bit confused when setting up my custom domain name for the generated API. As per the instructions in the readme, I was looking for my Gateway API endpoint ending with `.cloudfront.net`. The instructions may be clear for the more experienced AWS users but it might be a bit confusing for novice developers. 

P.S.: Thank you for the awesome script! I am very happy that I could shut down my EC2 bitwarden instance and have everything in a serverless form.